Refactors SettingsTab into new file
